% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{tokyor98_date}
\alias{tokyor98_date}
\title{Event Date of the 98th TokyoR}
\usage{
tokyor98_date()
}
\value{
A string object of the date
}
\description{
This function returns event date of the 98th TokyoR.
Date format is 'YYYY-MM-dd' by default
}
