% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{tokyor98_LT}
\alias{tokyor98_LT}
\title{Lightning Talk information of the 98th TokyoR}
\usage{
tokyor98_LT()
}
\value{
A data.frame object of LT presenters and each titles
}
\description{
This function returns basic information of LT pf the 98th TokyoR
}
\examples{
\dontrun{
  tokyor98_LT()
}

}
