% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TokyoR98demo.R
\docType{package}
\name{TokyoR98demo}
\alias{TokyoR98demo}
\alias{TokyoR98demo-package}
\title{TokyoR98demo: A package for providing basic information of the 98th TokyoR conference.}
\description{
This package was created to showcase how R packages are made through a short presentation by Koki Ando.
}
\author{
Koki Ando \email{koki.25.ando@gmail.com}
}
